04_everything_is_an_api/08_fastapi_todo/CRUD_FastAPI/pyproject.toml
[tool.poetry]
name = "crud-fastapi"
version = "0.1.0"
description = ""
authors = ["Shawaal Nadeem <110089331+Shawaal-Nadeem@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
uvicorn = "^0.28.0"
streamlit = "^1.32.2"
requests = "^2.31.0"
pandas = "^2.2.1"
pytest = "^8.1.1"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
